moved login.manager from routes to init
