+++ webapp/__init__.py
@@ -1,6 +1,8 @@
 from flask import Flask, render_template, flash, redirect, url_for
 from flask_sqlalchemy import SQLAlchemy
 from flask_migrate import Migrate, migrate
+from flask_login import LoginManager
+
 
 from webapp.config import Config
 
@@ -12,4 +14,13 @@ app.config.from_object(Config)
 db.init_app(app)  #  bounding app and bd
 migrate.init_app(app, db)  # bounding app, bd and migration instance
 
+login_manager = LoginManager()
+login_manager.init_app(app)
+login_manager.login_view = 'login'
+login_manager.login_message = "Войдите, чтобы просмотреть страницу"
+@login_manager.user_loader
+def load_user(user_id):
+    return User.query.get(user_id)
+
 from webapp import routes, model
+from webapp.model import db, User
